Added home page with links to admin and manager panels.  (#41)
* Added home page with links to admin and manager panels. Reorganized components accordingly

* Stylelint fix
