+++ front-end/src/SampleJson/card_colors.json
