+++ front-end/src/components/Admin/Header/Menu/index.js
@@ -1,0 +1,14 @@
+import React, {Component} from 'react';
+import './index.scss';
+
+class Menu extends Component {
+  render () {
+    return (
+      <div className="menu">
+        Tabs will be here.
+      </div>
+    );
+  }
+}
+
+export default Menu;
