+++ front-end/src/components/Admin/Main/GridEntities.js
@@ -1,9 +1,9 @@
-import packages from './SampleJson/packages.json';
-import services from './SampleJson/services.json';
-import serviceCategories from './SampleJson/service_categories.json';
-import serviceRules from './SampleJson/service_rules.json';
-import contracts from './SampleJson/contracts.json';
-import organizations from './SampleJson/organizations.json';
+import packages from '../../../SampleJson/packages.json';
+import services from '../../../SampleJson/services.json';
+import serviceCategories from '../../../SampleJson/service_categories.json';
+import serviceRules from '../../../SampleJson/service_rules.json';
+import contracts from '../../../SampleJson/contracts.json';
+import organizations from '../../../SampleJson/organizations.json';
 
 const GridEntities = [
   {
