+++ front-end/src/components/Admin/Main/index.scss
@@ -1,4 +1,4 @@
-.configurator {
+.main {
   display: grid;
   grid-template-columns: 1fr 3fr;
   grid-column-gap: .5vh;
