+++ front-end/src/components/Admin/index.scss
