--- front-end/src/components/Footer/index.js
@@ -1,16 +1,0 @@
-import React, {Component} from 'react';
-import './index.scss';
-
-class Footer extends Component {
-  render () {
-    return (
-      <footer className="footer">
-        <div className="footer__container">
-
-        </div>
-      </footer>
-    );
-  }
-}
-
-export default Footer;
