--- front-end/src/components/Footer/index.scss
@@ -1,15 +1,0 @@
-.footer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  grid-row: 3 / 4;
-  padding: 0 20px;
-  background-color: whitesmoke;
-
-  &__container {
-    display: flex;
-    align-items: center;
-    width: 100%;
-    height: 100%;
-  }
-}
