+++ front-end/src/components/Home/index.scss
@@ -1,0 +1,13 @@
+.home {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  grid-row: 2 / 3;
+  background-color: whitesmoke;
+
+  &__listitem {
+    margin: 5px 10px;
+    padding: 5px 10px;
+  }
+}
