--- front-end/src/components/Main/index.js
@@ -1,18 +1,0 @@
-import React, {Component} from 'react';
-import './index.scss';
-import {Route} from 'react-router-dom';
-import Home from './Home';
-import Configurator from './Configurator';
-
-class Main extends Component {
-  render () {
-    return (
-      <main className="main">
-        <Route exact path="/" component={Home} />
-        <Route path="/configurator" component={Configurator} />
-      </main>
-    );
-  }
-}
-
-export default Main;
