+++ front-end/src/components/Manager/index.js
@@ -1,0 +1,14 @@
+import React, {Component} from 'react';
+import './index.scss';
+
+class Menu extends Component {
+  render () {
+    return (
+      <div className="manager">
+        This is the home manager page.
+      </div>
+    );
+  }
+}
+
+export default Menu;
